SHINDIG-1897 - GadgetContext and SecurityToken have different definitions of module ID Committed for Marshall Shi
git-svn-id: https://svn.apache.org/repos/asf/shindig/trunk@1441000 13f79535-47bb-0310-9956-ffa450edef68

+++ java/gadgets/src/main/java/org/apache/shindig/gadgets/GadgetContext.java
@@ -60,7 +60,7 @@ public class GadgetContext {
   /**
    * @return The module id for this request.
    */
-  public int getModuleId() {
+  public long getModuleId() {
     return delegate == null ? 0 : delegate.getModuleId();
   }
